Make benchmark_for use 64-bit ints like the other benchmarks

+++ benchmark_for.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main() {
-    int sum = 0;
-    for (int i = 0; i < 100000000; i++) {
+    long long sum = 0;
+    for (long long i = 0; i < 100000000; i++) {
         sum += i;
     }
-    return sum;
+    return (int)sum;
 }
